adds handshake listener and moves client handler
